requirements.txt
astroid==2.12.13
autopep8==2.0.1
click==8.1.3
dill==0.3.6
grpcio==1.51.1
grpcio-tools==1.51.1
isort==5.11.4
lazy-object-proxy==1.8.0
mccabe==0.7.0
mypy==0.991
mypy-extensions==0.4.3
platformdirs==2.6.2
protobuf==4.21.12
pycodestyle==2.10.0
pylint==2.15.9
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
wrapt==1.14.1
